server/python_pipeline/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pandas==2.2.1
scikit-learn==1.4.1.post1
spacy==3.7.4
beautifulsoup4==4.12.3
requests==2.31.0
supabase==2.3.6
python-dotenv==1.0.1
